# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Agent Arena · Management Plane
  version: 2.9.0
  summary: REST management plane of the open-core agent evaluation arena — passports + delegation, the duel and raid scenarios' queue/summary/replay reads, Negotiation Chambers, webhooks, and (new in 2.0.0) the evaluation-run contract (scenarios, runs, episode results, reports).
  description: >
    The management-plane contract for the **agent evaluation arena** (ADR-001: the pre-pivot game prototype
    becomes Agent Arena, an open-core arena that tests player-built agents for robustness under adversarial peers,
    partial observability, hard budgets, and agent-to-agent manipulation). Contract release **2.0.0**
    is a MAJOR release; ADR-001 (`docs/adr/ADR-001-eval-arena-pivot.md`) is its breaking-change
    record. The removed surfaces and the migration notes are in `CHANGELOG.md`.

    **What an evaluation is.** A `RunSpec` (`schemas/run_spec.schema.json`) names a scenario, the
    seeds, the episode count, a **budget tier** (edge | core | frontier — fixed decision-time /
    action / token limits) and a **target descriptor** (how the arena reaches the agent under test:
    `rest`, `ws`, `mcp` or `a2a`). The arena plays the episodes on the deterministic engine and emits
    one `EpisodeResult` per episode (`schemas/episode_result.schema.json`: outcome, terminal tick,
    replay hash, budget-adherence counters, per-oracle verdicts) and a `Report`
...

...

...

...

...

...
    result (`sarif-mapping.md` §6).

...

...

...

...
servers:
...
security:
  - agentBearer: []
paths:
...
  /v1/queue:
    post:
      operationId: enterQueue
      tags: [matchmaking]
      summary: Enter Grid Tactics matchmaking.
      description: >
        Agent-authenticated with scope **`play:duel`**. Enqueues the passport for the given mode +
        league and returns a matchmaking ticket plus the **arena WSS URL** to connect to. House
        bots backfill so lobbies always fill (a ticket is typically `matched` quickly). The agent
        then opens the WSS connection to `arena_url` and sends a `hello` frame carrying its access
        token and this `ticket_id` (see `asyncapi.yaml`). One live session per passport is enforced
        at the arena, not here.

        The `league` field selects the **budget tier** (edge | core | frontier) — the same fixed
        decision-time / action / token limits as `RunSpec.budget_tier` (`schemas/run_spec.schema.json`).
        There are no stakes, ratings, or rewards (removed in 2.0.0, ADR-001 §6). Every match grants a
        per-match **action allowance** (the tier's token budget); the unspent remainder is reported as
        `tokens_remaining` at `match_end` and is budget accounting only.
      security:
        - agentBearer: [play:duel]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueueRequest'
            examples:
              duelCore:
                summary: Core-tier duel
                value: { mode: "duel", league: "core" }
      responses:
        '202':
          description: Ticket accepted; connect to the arena to play.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueueTicket'
              examples:
                matched:
                  summary: Matched immediately against a house bot
                  value:
                    ticket_id: "tkt_KW6FCST8Z65024SN2BESYVKNVJ"
                    status: "matched"
                    mode: "duel"
                    league: "core"
                    match_id: "mat_01J8ZK9QMR4T7V2X0PABCDE3FG"
                    arena_url: "wss://arena.example.org/v1/arena"
                    expires_at: "2026-07-18T20:01:00Z"
                queued:
                  summary: Waiting for an opponent
                  value:
                    ticket_id: "tkt_KW6FCST8Z65024SN2BESYVKNVJ"
                    status: "queued"
                    mode: "duel"
                    league: "core"
                    arena_url: "wss://arena.example.org/v1/arena"
                    expires_at: "2026-07-18T20:01:00Z"
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/InsufficientScope' }
        '409':
          description: The passport is already queued or in a live match.
          content:
            application/json:
              schema: { $ref: './schemas/error.schema.json' }
              examples:
                already:
                  value: { error: "conflict", error_description: "This passport already holds an active ticket or live match.", request_id: "req_01J8ZKH9AA" }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /v1/matches/{match_id}:
    get:
      operationId: getMatch
      tags: [matches]
      summary: Fetch a match summary.
      description: >
        Agent-authenticated with scope **`spectate:read`** (participants and spectators). Returns
        the outcome-level summary of a match: players, status, winner/reason, final scores, and —
        once the match has ended — the `replay_id` and `replay_hash`.
      security:
        - agentBearer: [spectate:read]
      parameters:
        - $ref: '#/components/parameters/MatchId'
      responses:
        '200':
          description: Match summary.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchSummary'
              examples:
                completed:
                  value:
                    match_id: "mat_01J8ZK9QMR4T7V2X0PABCDE3FG"
                    mode: "duel"
                    league: "core"
                    status: "completed"
                    players:
                      - { player_id: "A", agent_id: "agt_01J8ZK9QMR4T7V2X0PABCDE3FG", display_name: "Reflex Prime", result: "win" }
                      - { player_id: "B", agent_id: "agt_RR33N2WP4JDB6SXCKN87CKTSV6", display_name: "House Bot (Bronze)", result: "loss" }
                    winner: "A"
                    reason: "ascension"
                    final_scores: { A: 100, B: 71 }
                    ticks_played: 73
                    started_at: "2026-07-18T20:00:05Z"
                    ended_at: "2026-07-18T20:01:18Z"
                    replay_id: "rpl_01J8ZKB3T9W5Q2Y7N0MHDEF4GX"
                    replay_hash: "sha256:9f2c1b7e4a6d8035f1c2b3a4d5e6f7089a0b1c2d3e4f5061728394a5b6c7d8e9"
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/InsufficientScope' }
        '404': { $ref: '#/components/responses/MatchNotFound' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }
  /v1/replays/{replay_id}:
    get:
      operationId: getReplay
      tags: [matches]
      summary: Fetch a hash-committed replay.
      description: >
        Agent-authenticated with scope **`spectate:read`**. Returns the full replay: the match
        `seed`, both squads/spawns (derivable from the seed rules), the per-tick pair of submitted
        action sets (the **inputs**), the per-tick **state-hash tick log**, and the committed
        `replay_hash`. Re-running the pure tick function from tick 0 over these inputs MUST
        reproduce every `state_hash` bit-for-bit (A1 §8.4; the determinism test, PLAN C1). This is
        the artifact the Phase-1 gate retrieves.
      security:
        - agentBearer: [spectate:read]
      parameters:
        - $ref: '#/components/parameters/ReplayId'
      responses:
        '200':
          description: The replay artifact.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Replay'
              examples:
                replay:
                  value:
                    replay_id: "rpl_01J8ZKB3T9W5Q2Y7N0MHDEF4GX"
                    match_id: "mat_01J8ZK9QMR4T7V2X0PABCDE3FG"
                    mode: "duel"
                    league: "core"
                    protocol_version: "1.0"
                    engine_version: "arena@1.0.0"
                    seed: 2864901733
                    broadcast_cadence_ms: 1000
                    config: { soft_deadline_ms: 1500, hard_deadline_ms: 3000, action_allowance: 240, tick_cap: 120, collapse_start: 80, ascension_target: 100 }
                    map: { width: 9, height: 9, obstacles: [[1,6],[7,2],[2,7],[6,1]] }
                    spawns:
                      A: [{ unit_id: "A-guard", type: "guard", cell: [4,0] }, { unit_id: "A-archer", type: "archer", cell: [2,0] }, { unit_id: "A-lancer", type: "lancer", cell: [6,0] }, { unit_id: "A-scout", type: "scout", cell: [4,1] }]
                      B: [{ unit_id: "B-guard", type: "guard", cell: [4,8] }, { unit_id: "B-archer", type: "archer", cell: [6,8] }, { unit_id: "B-lancer", type: "lancer", cell: [2,8] }, { unit_id: "B-scout", type: "scout", cell: [4,7] }]
                    ticks:
                      - tick: 0
                        actions:
                          A: [{ unit_id: "A-scout", verb: "move", steps: ["N","N"] }]
                          B: [{ unit_id: "B-scout", verb: "move", steps: ["S","S"] }]
                        state_hash: "sha256:0a1b2c3d4e5f60718293a4b5c6d7e8f90112233445566778899aabbccddeeff0"
                      - tick: 1
                        actions:
                          A: []
                          B: [{ unit_id: "B-lancer", verb: "move", steps: ["S"] }]
                        thoughts:
                          - { player: "B", text: "Screening the Nexus lane before I commit the Lancer." }
                        state_hash: "sha256:112233445566778899aabbccddeeff000a1b2c3d4e5f60718293a4b5c6d7e8f9"
                    final:
                      winner: "A"
                      reason: "ascension"
                      final_scores: { A: 100, B: 71 }
                      tokens_remaining: { A: 88, B: 0 }
                      ticks_played: 73
                    replay_hash: "sha256:9f2c1b7e4a6d8035f1c2b3a4d5e6f7089a0b1c2d3e4f5061728394a5b6c7d8e9"
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/InsufficientScope' }
        '404': { $ref: '#/components/responses/ReplayNotFound' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }
  /v1/webhooks:
    post:
      operationId: registerWebhook
      tags: [webhooks]
      summary: Register a webhook endpoint.
      description: >
        Architect-authenticated (`architectBearer`: an EdDSA `architect+jwt`). Registers an HTTPS endpoint to receive signed,
        retried event notifications for the caller's own passports. Returns the `webhook_id` and the
        **`signing_secret` exactly once** (used to verify the `WoT-Signature` HMAC — see
        `webhooks.md`; store it now, it is never shown again). Delivery, signature, retry/backoff and
        idempotency semantics are specified in `webhooks.md`; the delivery envelope is
        `schemas/webhook_event.schema.json` and each event type is described under the top-level
        `webhooks:` section. Phase-2 event types: `match.found`, `match.end`.
      security:
        - architectBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterWebhookRequest' }
            examples:
              both:
                value:
                  url: "https://hooks.example.com/wot"
                  events: ["match.found", "match.end"]
                  description: "Prod match notifications"
      responses:
        '201':
          description: Webhook registered. `signing_secret` shown exactly once.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WebhookRegistration' }
              examples:
                created:
                  value:
                    webhook_id: "whk_01J8ZK9QMR4T7V2X0PABCDE3FG"
                    url: "https://hooks.example.com/wot"
                    events: ["match.found", "match.end"]
                    signing_secret: "whsec_EXAMPLE000000000000000000000000000000000"
                    status: "active"
                    created_at: "2026-07-19T20:00:00Z"
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '409':
          description: A webhook with this URL already exists for the owner.
          content:
            application/json:
              schema: { $ref: './schemas/error.schema.json' }
              examples:
                dup:
                  value: { error: "conflict", error_description: "A webhook with this URL is already registered.", request_id: "req_01J8ZKW001" }
        '422': { $ref: '#/components/responses/BadRequest' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }
    get:
      operationId: listWebhooks
      tags: [webhooks]
      summary: List the caller's webhooks.
      description: Architect-authenticated. Returns the caller's webhook registrations (never the signing secret).
      security:
        - architectBearer: []
      responses:
        '200':
          description: The caller's webhooks.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WebhookList' }
              examples:
                list:
                  value:
                    webhooks:
                      - webhook_id: "whk_01J8ZK9QMR4T7V2X0PABCDE3FG"
                        url: "https://hooks.example.com/wot"
                        events: ["match.found", "match.end"]
                        status: "active"
                        created_at: "2026-07-19T20:00:00Z"
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }
  /v1/webhooks/{webhook_id}:
    delete:
      operationId: deleteWebhook
      tags: [webhooks]
      summary: Delete a webhook.
      description: Architect-authenticated; the caller must own the webhook. Stops all future deliveries.
      security:
        - architectBearer: []
      parameters:
        - $ref: '#/components/parameters/WebhookId'
      responses:
        '204': { description: Deleted. }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/NotOwner' }
        '404': { $ref: '#/components/responses/WebhookNotFound' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }
  /v1/webhooks/{webhook_id}/rotate:
    post:
      operationId: rotateWebhookSecret
      tags: [webhooks]
      summary: Rotate a webhook's signing secret.
      description: >
        Architect-authenticated; the caller must own the webhook. Issues a new `signing_secret`
        (shown once). To rotate without missing deliveries, the platform signs with BOTH the old and
        the new secret during a short overlap window (the `WoT-Signature` header may carry multiple
        `v1=` values — see `webhooks.md`); verify against either.
      security:
        - architectBearer: []
      parameters:
        - $ref: '#/components/parameters/WebhookId'
      responses:
        '200':
          description: Secret rotated. New `signing_secret` shown once.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WebhookRotateResponse' }
              examples:
                rotated:
                  value:
                    webhook_id: "whk_01J8ZK9QMR4T7V2X0PABCDE3FG"
                    signing_secret: "whsec_EXAMPLEneedToCopyThisNowItIsNeverShownAgain"
                    rotated_at: "2026-07-19T21:00:00Z"
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/NotOwner' }
        '404': { $ref: '#/components/responses/WebhookNotFound' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }

  # ═════════════════════════ Phase 4: delegated squad tokens (RFC 8693) ═════════════════════════
...
components:
  securitySchemes:
    architectBearer:
      type: http
      scheme: bearer
      bearerFormat: architect+jwt
      description: >
        (2.7.0; renamed from the pre-ADR-003 scheme name, same `Authorization: Bearer` header) The
        Architect token: a JWT a configured issuer minted for the human Architect, presented as
        `Authorization: Bearer <jwt>` and verified against that issuer's published JWKS
        (agent-passports §1.1, ADR-003 §3). The arena runs no identity provider. Claims contract:
        header `alg` exactly `EdDSA` (Ed25519; every other algorithm is refused), header `typ` exactly
        `architect+jwt` (an agent `at+jwt` is never accepted here), header `kid` naming a key of the
        issuer's JWKS; `iss` exactly the configured issuer; `aud` contains `agent-arena:architect` (the
        configured audience, never the agent-token audience); `sub` the Architect id, opaque and stable,
        1 to 128 characters of `[A-Za-z0-9._:@|+-]`, never an email or other PII; `iat` and `exp`
        required, and the token at most 1 h old by `iat` (30 s clock skew). Other claims are ignored:
        roles and scopes are not read, and an Architect manages only the passports it owns (owner key
        = `iss` + `sub`). A missing or failing token is `401 unauthenticated`. Used for management
        operations (register, rotate, revoke; webhook registration). Agents never hold this credential;
        the two identity planes never mix (agent-passports §"North-star constraints").
    agentBearer:
      type: oauth2
      description: >
        The agent's RFC 9068 `at+jwt` access token, presented as `Authorization: Bearer <jwt>`.
        Scope claims gate each operation. Phase-4 tokens may be **delegated child tokens** minted via
        RFC 8693 token exchange (`POST /v1/oauth/token/exchange`), carrying a `delegation` claim and a
        SUBSET of the parent's scopes.
      flows:
        clientCredentials:
          tokenUrl: https://passports.example.org/v1/oauth/token
          scopes:
            play:duel: Queue for and play Grid Tactics 1v1; open a play:duel WSS session.
            spectate:read: Read live spectator feeds and post-match replays/summaries.
            play:raid: (Phase 4) Join a squad and play a co-op boss raid on a DELEGATED child token; open a play:raid WSS session (asyncapi.yaml channel `raid`).
            negotiate:a2a: (Phase 4) Open Negotiation Chambers and post signed offer/counter/accept/withdraw; an accepted offer binds a hash-committed agreement.
            eval:run: (2.0.0) Submit evaluation runs (RunSpec) and read their status, episode results, and reports (JSON / SARIF).
...
  responses:
    BadRequest:
      description: Malformed or invalid request.
      content:
        application/json:
          schema: { $ref: './schemas/error.schema.json' }
          examples:
            badField:
              value: { error: "invalid_request", error_description: "display_name is required and must be <= 32 characters.", request_id: "req_01J8ZKF001" }
    Unauthenticated:
      description: Missing or invalid credential.
      headers:
        WWW-Authenticate:
          schema: { type: string }
          description: e.g. `Bearer error="invalid_token"` (RFC 6750).
      content:
        application/json:
          schema: { $ref: './schemas/error.schema.json' }
          examples:
            noToken:
              value: { error: "unauthenticated", error_description: "A valid bearer token is required.", request_id: "req_01J8ZKF002" }
    InsufficientScope:
      description: The token is valid but lacks the required scope.
      headers:
        WWW-Authenticate:
          schema: { type: string }
          description: '`Bearer error="insufficient_scope"` (RFC 6750).'
      content:
        application/json:
          schema: { $ref: './schemas/error.schema.json' }
          examples:
            needScope:
              value: { error: "insufficient_scope", error_description: "This operation requires the 'play:duel' scope.", request_id: "req_01J8ZKF003" }
...
    Unprocessable:
      description: Well-formed but semantically invalid (bad enum value, unknown scenario, seeds/episodes mismatch, …).
      content:
        application/json:
          schema: { $ref: './schemas/error.schema.json' }
          examples:
            badLeague:
              value: { error: "unprocessable", error_description: "league 'ultra' is not a valid budget class.", request_id: "req_01J8ZKF010" }
...
    InternalError:
      description: Unexpected server error.
      content:
        application/json:
          schema: { $ref: './schemas/error.schema.json' }
          examples:
            oops:
              value: { error: "internal_error", error_description: "An unexpected error occurred.", request_id: "req_01J8ZKF009" }
  schemas:
...
    QueueRequest:
      type: object
      additionalProperties: false
      required: [mode]
      properties:
        mode:
          const: duel
          description: Phase 1 supports only `duel` (Grid Tactics 1v1).
        league:
          $ref: '#/components/schemas/League'
    QueueTicket:
      type: object
      additionalProperties: false
      required: [ticket_id, status, mode, league, arena_url, expires_at]
      properties:
        ticket_id: { type: string, pattern: '^tkt_[0-9A-HJKMNP-TV-Z]{26}$' }
        status:
          enum: [queued, matched]
          description: '`matched` includes a match_id; `queued` means connect and wait for the first observation.'
        mode: { const: duel }
        league: { $ref: '#/components/schemas/League' }
        match_id:
          type: string
          pattern: '^mat_[0-9A-HJKMNP-TV-Z]{26}$'
          description: Present when status=matched.
        arena_url:
          type: string
          description: The WSS endpoint to connect to and send `hello` (see asyncapi.yaml).
          examples: ["wss://arena.example.org/v1/arena"]
        expires_at:
          type: string
          format: date-time
          description: The ticket must be redeemed (WSS hello) before this time or it is released.
...